Move evaluteResult into eval util
