+++ metricproviders/prometheus/prometheus_test.go
@@ -171,95 +171,6 @@ func TestGarbageCollect(t *testing.T) {
 	assert.NoError(t, err)
 }
 
-func TestEvaluateResultWithSuccess(t *testing.T) {
-	p := Provider{}
-	metric := v1alpha1.Metric{
-		SuccessCondition: "true",
-		FailureCondition: "false",
-	}
-	status := p.evaluateResult(true, metric)
-	assert.Equal(t, v1alpha1.AnalysisPhaseSuccessful, status)
-}
-
-func TestEvaluateResultWithFailure(t *testing.T) {
-	p := Provider{}
-	metric := v1alpha1.Metric{
-		SuccessCondition: "true",
-		FailureCondition: "true",
-	}
-	status := p.evaluateResult(true, metric)
-	assert.Equal(t, v1alpha1.AnalysisPhaseFailed, status)
-
-}
-
-func TestEvaluateResultInconclusive(t *testing.T) {
-	p := Provider{}
-	metric := v1alpha1.Metric{
-		SuccessCondition: "false",
-		FailureCondition: "false",
-	}
-	status := p.evaluateResult(true, metric)
-	assert.Equal(t, v1alpha1.AnalysisPhaseInconclusive, status)
-}
-
-func TestEvaluateResultNoSuccessConditionAndNotFailing(t *testing.T) {
-	p := Provider{}
-	metric := v1alpha1.Metric{
-		SuccessCondition: "",
-		FailureCondition: "false",
-	}
-	status := p.evaluateResult(true, metric)
-	assert.Equal(t, v1alpha1.AnalysisPhaseSuccessful, status)
-}
-
-func TestEvaluateResultNoFailureConditionAndNotSuccessful(t *testing.T) {
-	p := Provider{}
-	metric := v1alpha1.Metric{
-		SuccessCondition: "false",
-		FailureCondition: "",
-	}
-	status := p.evaluateResult(true, metric)
-	assert.Equal(t, v1alpha1.AnalysisPhaseFailed, status)
-}
-
-func TestEvaluateResultNoFailureConditionAndNoSuccessCondition(t *testing.T) {
-	p := Provider{}
-	metric := v1alpha1.Metric{
-		SuccessCondition: "",
-		FailureCondition: "",
-	}
-	status := p.evaluateResult(true, metric)
-	assert.Equal(t, v1alpha1.AnalysisPhaseSuccessful, status)
-}
-
-func TestEvaluateResultWithErrorOnSuccessCondition(t *testing.T) {
-	logCtx := log.WithField("test", "test")
-	p := Provider{
-		logCtx: *logCtx,
-	}
-	metric := v1alpha1.Metric{
-		SuccessCondition: "a == true",
-		FailureCondition: "true",
-	}
-	status := p.evaluateResult(true, metric)
-	assert.Equal(t, v1alpha1.AnalysisPhaseError, status)
-
-}
-
-func TestEvaluateResultWithErrorOnFailureCondition(t *testing.T) {
-	logCtx := log.WithField("test", "test")
-	p := Provider{
-		logCtx: *logCtx,
-	}
-	metric := v1alpha1.Metric{
-		SuccessCondition: "true",
-		FailureCondition: "a == true",
-	}
-	status := p.evaluateResult(true, metric)
-	assert.Equal(t, v1alpha1.AnalysisPhaseError, status)
-
-}
-
 func TestProcessScalarResponse(t *testing.T) {
 	logCtx := log.WithField("test", "test")
 	p := Provider{
