+++ metricproviders/webmetric/webmetric.go
@@ -100,51 +100,8 @@ func (p *Provider) parseResponse(metric v1alpha1.Metric, response *http.Response
 	}
 	out := buf.String()
 
-	status := p.evaluateResponse(metric, out)
+	status := evaluate.EvaluateResult(out, metric, p.logCtx)
 	return out, status, nil
-}
-
-func (p *Provider) evaluateResponse(metric v1alpha1.Metric, result interface{}) v1alpha1.AnalysisPhase {
-	successCondition := false
-	failCondition := false
-	var err error
-
-	if metric.SuccessCondition != "" {
-		successCondition, err = evaluate.EvalCondition(result, metric.SuccessCondition)
-		if err != nil {
-			p.logCtx.Warning(err.Error())
-			return v1alpha1.AnalysisPhaseError
-		}
-	}
-	if metric.FailureCondition != "" {
-		failCondition, err = evaluate.EvalCondition(result, metric.FailureCondition)
-		if err != nil {
-			return v1alpha1.AnalysisPhaseError
-		}
-	}
-
-	switch {
-	case metric.SuccessCondition == "" && metric.FailureCondition == "":
-		//Always return success unless there is an error
-		return v1alpha1.AnalysisPhaseSuccessful
-	case metric.SuccessCondition != "" && metric.FailureCondition == "":
-		// Without a failure condition, a measurement is considered a failure if the measurement's success condition is not true
-		failCondition = !successCondition
-	case metric.SuccessCondition == "" && metric.FailureCondition != "":
-		// Without a success condition, a measurement is considered a successful if the measurement's failure condition is not true
-		successCondition = !failCondition
-	}
-
-	if failCondition {
-		return v1alpha1.AnalysisPhaseFailed
-	}
-
-	if !failCondition && !successCondition {
-		return v1alpha1.AnalysisPhaseInconclusive
-	}
-
-	// If we reach this code path, failCondition is false and successCondition is true
-	return v1alpha1.AnalysisPhaseSuccessful
 }
 
 // Resume should not be used the WebMetric provider since all the work should occur in the Run method
